refactor(trip-controller-test): Extract generator utils
