+++ server/src/controllers/__tests__/trip.controller.test.ts
@@ -2,47 +2,8 @@ import mockingoose from 'mockingoose';
 import mongoose from 'mongoose';
 import * as tripController from './../trip.controller';
 import Trip from './../../models/trip.model';
-
-import faker from 'faker';
-import dayjs from 'dayjs';
-import { hasUncaughtExceptionCaptureCallback } from 'process';
-
-function buildTrip(): any {
-  return {
-    _id: mongoose.Types.ObjectId(),
-    name: faker.lorem.words(),
-    country: faker.address.country(),
-    startDate: transformTimestampToDate(faker.date.future()),
-    endDate: transformTimestampToDate(faker.date.future()),
-    stops: [faker.address.city(), faker.address.city()],
-  };
-}
-
-function transformTimestampToDate(timestamp: Date): Date {
-  return new Date(dayjs().format('YYYY-MM-DD'));
-}
-
-function buildReq(overrides = {}) {
-  const req: any = {
-    body: {},
-    params: {},
-    ...overrides,
-  };
-  return req;
-}
-
-function buildRes(overrides = {}) {
-  const res: any = {
-    json: jest.fn(() => res).mockName('json'),
-    status: jest.fn(() => res).mockName('status'),
-    ...overrides,
-  };
-  return res;
-}
-
-function jsonStrinigifyParse(item: any) {
-  return JSON.parse(JSON.stringify(item));
-}
+import { buildTrip, buildReq, buildRes } from './../../../test/utils/generate';
+import { jsonStrinigifyParse } from './../../../test/utils/helpers';
 
 beforeEach(() => {
   mockingoose.resetAll();
