+++ server/test/utils/helpers.ts
@@ -1,0 +1,3 @@
+export function jsonStrinigifyParse(item: any) {
+  return JSON.parse(JSON.stringify(item));
+}
